+++ NEWS.md
@@ -4,6 +4,7 @@ JuMP release notes
 Unversioned
 -----------
 
+  * Specify variable lower/upper bounds in ``@variable`` using the ``lowerbound`` and ``upperbound`` keyword arguments.
   * Change name printed for variable using the ``basename`` keyword argument to ``@defVar``.
 
 Version 0.12.2 (March 9, 2016)